requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-tenants==3.5.0
psycopg2-binary==2.9.7
redis==4.6.0
django-redis==5.3.0
celery==5.3.4
python-decouple==3.8
requests==2.31.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.0.1
boto3==1.29.7
django-storages==1.14.2
PyJWT==2.8.0
drf-yasg==1.21.7
qrcode==7.4.2
resend==2.23.0
pusher==3.3.2
